Segment.tcp_output_really: update the rcv_wnd of the tcp_state
This updated rcv_window has just been sent over the wire, but the internal state
(control_block) has not been updated. Some input validation (e.g. in_window)
rely on the control_block data -- so better keep them synchronised.

TODO: check with a fresh head the consequences, and whether there are other
occassions where this is necessary.

Also, the User.recv may need to send out data (i.e. a window update -- in case
the rcv_wnd was 0, it could now be bigger). So, the API has been refined to take
a Mtime.t as input, and an output option.