allow multiple segments to be sent out from tcp_output_perhaps
the API is now to return a list of segments instead of an option
this fixes #25

also, to avoid when a server is of the form:
 accept()
 write <my big chunk of data>
 close

then the write will emit any number of segments until the congestion window is
full. now, close won't fit their FIN into any segment, so all we can do is wait
for ack (or timer, which retransmits) from the other side before continuing
(then the congestion window opens again, and we proceed sending more data).

this change includes that User.close() and User.shutdown() do not set the
tf_shouldacknow -- in contrast to the HOL specification.