di3_datastugg_really: fix reassembly and fin
now that we have a reassembly queue, if we receive an out-of-order segment,
which is pushed to the queue and has fin flag, do not claim that fin is true.

once the segment is pulled from the queue, the fin will be back. handling the
fin prematurely leads to bad state transitions, which leads to confusion, and
eventually assertion failure.

this commit should fix #26, where I gathered a trace where first an out-of-order
fin was received, followed by some further bytes of data (retransmission).